fix methods according to security

+++ backend/src/main/java/itmo/is/dto/domain/request/AddPersonToGroupRequest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record AddPersonToGroupRequest(
         @JsonProperty(value = "group_id", required = true)
-        Long group_id,
+        Integer group_id,
 
         @JsonProperty(value = "person_id", required = true)
         Integer person_id
